Move db helpers to rocksdb crate
