+++ bin/strata-client/src/main.rs
@@ -25,8 +25,9 @@ use strata_eectl::engine::ExecEngineCtl;
 use strata_evmexec::{engine::RpcExecEngineCtl, EngineRpcClient};
 use strata_primitives::params::Params;
 use strata_rocksdb::{
-    broadcaster::db::BroadcastDb, open_rocksdb_database, sequencer::db::SequencerDB, DbOpsConfig,
-    RBSeqBlobDb,
+    broadcaster::db::BroadcastDb, init_broadcaster_database, init_core_dbs,
+    init_sequencer_database, open_rocksdb_database, sequencer::db::SequencerDB, CommonDb,
+    DbOpsConfig, RBSeqBlobDb,
 };
 use strata_rpc_api::{StrataAdminApiServer, StrataApiServer, StrataSequencerApiServer};
 use strata_status::StatusChannel;
